move to ui/qt/network
