+++ selfdrive/ui/qt/network/networking.h
@@ -2,7 +2,7 @@
 
 #include <vector>
 
-#include "selfdrive/ui/qt/offroad/wifiManager.h"
+#include "selfdrive/ui/qt/network/wifimanager.h"
 #include "selfdrive/ui/qt/widgets/input.h"
 #include "selfdrive/ui/qt/widgets/ssh_keys.h"
 #include "selfdrive/ui/qt/widgets/toggle.h"
