+++ selfdrive/ui/qt/network/wifimanager.cc
@@ -1,4 +1,4 @@
-#include "selfdrive/ui/qt/offroad/wifiManager.h"
+#include "selfdrive/ui/qt/network/wifimanager.h"
 
 #include "selfdrive/ui/ui.h"
 #include "selfdrive/ui/qt/widgets/prime.h"
